requirements.txt
celery==5.3.4
pandas==1.1.5
psycopg2-binary==2.8.6
python-dotenv==0.17.1
fastapi==0.65.1
uvicorn==0.14.0